requirements.txt
# Core ML & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# AutoML Frameworks
autogluon>=1.0.0
# pycaret>=3.0.0  # Backup - install if needed
optuna>=3.0.0

# EDA & Visualization
ydata-profiling>=4.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Processing
imbalanced-learn>=0.11.0
category-encoders>=2.6.0
featuretools>=1.28.0

# Model Interpretation
shap>=0.43.0
# lime>=0.2.0  # Optional

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0

# Reporting
jinja2>=3.1.0
jupyter>=1.0.0
nbformat>=5.9.0
nbconvert>=7.8.0

# API Integration (for AI insights)
anthropic>=0.5.0  # Claude API

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
